+++ dev-test/gql-tag-operations/graphql/gql.ts
@@ -13,7 +13,12 @@ import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/
  * Therefore it is highly recommended to use the babel or swc plugin for production.
  * Learn more about it here: https://the-guild.dev/graphql/codegen/plugins/presets/preset-client#reducing-bundle-size
  */
-const documents = {
+type Documents = {
+  '\n  query Foo {\n    Tweets {\n      id\n    }\n  }\n': typeof types.FooDocument;
+  '\n  fragment Lel on Tweet {\n    id\n    body\n  }\n': typeof types.LelFragmentDoc;
+  '\n  query Bar {\n    Tweets {\n      ...Lel\n    }\n  }\n': typeof types.BarDocument;
+};
+const documents: Documents = {
   '\n  query Foo {\n    Tweets {\n      id\n    }\n  }\n': types.FooDocument,
   '\n  fragment Lel on Tweet {\n    id\n    body\n  }\n': types.LelFragmentDoc,
   '\n  query Bar {\n    Tweets {\n      ...Lel\n    }\n  }\n': types.BarDocument,