+++ examples/persisted-documents-string-mode/src/gql/gql.ts
@@ -12,7 +12,10 @@ import * as types from './graphql';
  * Therefore it is highly recommended to use the babel or swc plugin for production.
  * Learn more about it here: https://the-guild.dev/graphql/codegen/plugins/presets/preset-client#reducing-bundle-size
  */
-const documents = {
+type Documents = {
+  '\n  query HelloQuery {\n    hello\n  }\n': typeof types.HelloQueryDocument;
+};
+const documents: Documents = {
   '\n  query HelloQuery {\n    hello\n  }\n': types.HelloQueryDocument,
 };