+++ examples/react/http-executor/src/App.tsx
@@ -5,7 +5,7 @@ import Film from './Film';
 import { graphql, DocumentType } from './gql';
 
 const executor = buildHTTPExecutor({
-  endpoint: 'https://swapi-graphql.netlify.app/.netlify/functions/index',
+  endpoint: 'https://graphql.org/graphql/',
 });
 
 const AllFilmsWithVariablesQuery = graphql(/* GraphQL */ `
@@ -21,7 +21,7 @@ const AllFilmsWithVariablesQuery = graphql(/* GraphQL */ `
 `);
 
 // we could also define a client:
-// `const client = new GraphQLClient('https://swapi-graphql.netlify.app/.netlify/functions/index')`
+// `const client = new GraphQLClient('https://graphql.org/graphql/')`
 // and use:
 // `client.request(allFilmsWithVariablesQueryDocument, { first: 10 })`