+++ examples/typescript-esm/src/main.ts
@@ -32,7 +32,7 @@ const AllPeopleWithVariablesQueryDocument = graphql(/* GraphQL */ `
   }
 `);
 
-const apiUrl = 'https://swapi-graphql.netlify.app/.netlify/functions/index';
+const apiUrl = 'https://graphql.org/graphql/';
 
 executeOperation(apiUrl, AllPeopleQueryDocument).then(res => {
   if (res.errors) {