+++ examples/typescript-graphql-request/codegen.ts
@@ -2,7 +2,7 @@
 import { CodegenConfig } from '@graphql-codegen/cli';
 
 const config: CodegenConfig = {
-  schema: 'https://swapi-graphql.netlify.app/.netlify/functions/index',
+  schema: 'https://graphql.org/graphql/',
   documents: ['src/**/*.ts'],
   generates: {
     './src/gql/': {