+++ scripts/print-example-ci-command.js
@@ -6,11 +6,26 @@ const packageJSON = fg.sync(['examples/**/package.json'], { ignore: ['**/node_mo
 
 const ignoredPackages = [];
 
+const exampleTypeMap = {
+  all: 'all',
+  swc: 'swc',
+  normal: 'normal',
+};
+const exampleType = exampleTypeMap[process.env.EXAMPLE_TYPE] || 'all';
+
 const result = packageJSON.reduce(
   (res, packageJSONPath) => {
-    const { name } = fs.readJSONSync(packageJSONPath);
+    const { name, devDependencies } = fs.readJSONSync(packageJSONPath);
 
     if (ignoredPackages.includes(name)) {
+      res.ignored.push(name);
+      return res;
+    }
+
+    if (
+      (exampleType === 'swc' && !devDependencies['@graphql-codegen/client-preset-swc-plugin']) ||
+      (exampleType === 'normal' && devDependencies['@graphql-codegen/client-preset-swc-plugin'])
+    ) {
       res.ignored.push(name);
       return res;
     }