+++ packages/plugins/typescript/document-nodes/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @graphql-codegen/typescript-document-nodes
 
+## 4.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`60dd72f`](https://github.com/dotansimha/graphql-code-generator/commit/60dd72fb103fd7fd70b4e1def98da29588865517)]:
+  - @graphql-codegen/visitor-plugin-common@5.6.1
+
 ## 4.0.12
 
 ### Patch Changes