+++ packages/plugins/typescript/gql-tag-operations/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @graphql-codegen/gql-tag-operations
 
+## 4.0.14
+
+### Patch Changes
+
+- [#10192](https://github.com/dotansimha/graphql-code-generator/pull/10192) [`ec07018`](https://github.com/dotansimha/graphql-code-generator/commit/ec070189a1a3c4d41f2457b56a68b506c81f28ba) Thanks [@brianhuang822](https://github.com/brianhuang822)! - Have gql-tag-operations generate the type for document registry
+
+## 4.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`60dd72f`](https://github.com/dotansimha/graphql-code-generator/commit/60dd72fb103fd7fd70b4e1def98da29588865517)]:
+  - @graphql-codegen/visitor-plugin-common@5.6.1
+
 ## 4.0.12
 
 ### Patch Changes